hypershift upstream Fix 0.1
