+++ ocs_ci/deployment/helpers/hypershift_base.py
@@ -593,25 +593,6 @@ class HyperShiftBase:
         logger.info(f"Multicluster engine version: {cmd_res.stdout.decode('utf-8')}")
         return cmd_res.stdout.decode("utf-8")
 
-    def disable_multicluster_engine(self):
-        """
-        Disable multicluster engine on cluster
-        """
-        logger.info("Disabling multicluster engine")
-        cmd = (
-            "oc patch mce multiclusterengine "
-            '-p \'{"spec":{"overrides":{"components":['
-            '{"enabled":false, "name":"hypershift"},'
-            '{"enabled":false, "name":"hypershift-local-hosting"}, '
-            '{"enabled":false, "name":"local-cluster"}'
-            "]}}}' --type=merge"
-        )
-        cmd_res = exec_cmd(cmd, shell=True)
-        if cmd_res.returncode != 0:
-            logger.error(f"Failed to disable multicluster engine\n{cmd_res.stderr}")
-            return
-        logger.info(cmd_res.stdout.decode("utf-8").splitlines())
-
     def install_hypershift_upstream_on_cluster(self):
         """
         Install hypershift on the cluster
